Change Infra Env wizard to form
