--- src/components/InfraEnv/Wizard/index.ts
@@ -1,1 +1,0 @@
-export { default as InfraWizard } from './InfraWizard';
