+++ src/components/InfraEnv/index.ts
@@ -1,2 +1,2 @@
-export * from './Wizard';
+export { default as InfraEnvForm } from './InfraEnvForm';
 export { default as EnvironmentDetails } from './EnvironmentDetails';
